se envio el user y el mail a la base de datos y en ella se logra visualizar los datos
