+++ src/index.html
@@ -13,21 +13,21 @@
     <ul>
       <li><a class="active" href="#">Home</a></li>
       <li><a href="#">Iniciar Sesión</a></li>
-      <li><a href="#">Cerrar Sesión</a></li>
-      <li style="float:right"><a href="#about">Registrarme</a></li>
+      <li><a href="#" data-toggle="modal" data.target="signup">Cerrar Sesión</a></li>
+      <li style="float:right"><a href="about.html">Registrarme</a></li>
     </ul>
   </header>
 
 </head>
 <body>
 
-  <div class="container">
+  <div class="signup">
     <div class="form-background">
     <h1>Iniciar sesión</h1>
-   <form>
-    <input type="text" class="barra" name="user" placeholder="Usuario">
+   <form id="form-login">
+    <input type="text" class="barra" id="user" placeholder="Usuario">
     <br>
-    <input type="text" class="barra" name="Email" placeholder="Email">
+    <input type="text" class="barra" id="mail" placeholder="Email">
     <br><br>
     <input type="submit" value="Ingresar">
   </form>
@@ -46,6 +46,8 @@
   </div>
   </div>
  
+
+  
 
 <!-- Modal
 <div class="modal fade" id="signupModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
